requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.23.0
opencv-python>=4.7.0
tensorflow>=2.12.0
scikit-learn>=1.2.0
Pillow>=9.5.0
shap>=0.42.0

